+++ front/src/pages/settings/developers/api-keys/SettingsDevelopersApiKeyDetail.tsx
@@ -10,8 +10,8 @@ import { ApiKeyInput } from '@/settings/developers/components/ApiKeyInput';
 import { GET_API_KEYS } from '@/settings/developers/graphql/queries/getApiKeys';
 import { useGeneratedApiKeys } from '@/settings/developers/hooks/useGeneratedApiKeys';
 import { generatedApiKeyFamilyState } from '@/settings/developers/states/generatedApiKeyFamilyState';
-import { computeNewExpirationDate } from '@/settings/developers/utils.py/compute-new-expiration-date';
-import { formatExpiration } from '@/settings/developers/utils.py/format-expiration';
+import { computeNewExpirationDate } from '@/settings/developers/utils/compute-new-expiration-date';
+import { formatExpiration } from '@/settings/developers/utils/format-expiration';
 import { IconRepeat, IconSettings, IconTrash } from '@/ui/display/icon';
 import { H2Title } from '@/ui/display/typography/components/H2Title';
 import { Button } from '@/ui/input/button/components/Button';
